requirements.txt

Pygments==2.9.0
rubygemsrecipe==0.4.1
arabic-reshaper==2.1.3
base58==2.1.0
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
fpdf==1.7.2
gunicorn==20.1.0
Pillow==8.2.0
premailer==3.9.0
PyMySQL==1.0.2
PyPDF2==1.26.0
pypng==0.0.20
PyQRCode==1.2.1
python-bidi==0.4.2
pyzbar==0.1.8
qr==0.6.0
qrcode==7.1
qrtools==0.0.2
reportlab==3.5.68
reportlib==3.4.0
requests==2.25.1
requests-oauthlib==1.3.0
SQLAlchemy==1.4.20
uttlv==0.4.0

